Move datetime to separate file
