+++ src/assets/blocks/block-controls/block-controls.jsx
@@ -1,4 +1,4 @@
-import { AdvDateTimeControl } from "../0-adv-components/components.jsx";
+import { AdvDateTimeControl } from "../0-adv-components/datetime.jsx";
 
 (function ( wpI18n, wpHooks, wpBlocks, wpBlockEditor, wpComponents, wpCompose ) {
     wpBlockEditor = wp.blockEditor || wp.editor;
