+++ src/assets/blocks/blocks.js
@@ -5038,11 +5038,7 @@ module.exports = StyleToObject;
 Object.defineProperty(exports, "__esModule", {
     value: true
 });
-
-var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
-
 exports.AdvColorControl = AdvColorControl;
-exports.AdvDateTimeControl = AdvDateTimeControl;
 function AdvColorControl(props) {
     var _wp$components = wp.components,
         ColorIndicator = _wp$components.ColorIndicator,
@@ -5073,111 +5069,6 @@ function AdvColorControl(props) {
             value: value,
             onChange: onChange
         })
-    );
-}
-
-function AdvDateTimeControl(props) {
-    var _wp$components2 = wp.components,
-        Button = _wp$components2.Button,
-        DateTimePicker = _wp$components2.DateTimePicker,
-        Popover = _wp$components2.Popover,
-        Tooltip = _wp$components2.Tooltip;
-    var _wp$element = wp.element,
-        Fragment = _wp$element.Fragment,
-        useState = _wp$element.useState;
-    var __ = wp.i18n.__;
-
-    var _useState = useState(false),
-        _useState2 = _slicedToArray(_useState, 2),
-        popupState = _useState2[0],
-        setPopupState = _useState2[1];
-
-    var togglePopup = function togglePopup() {
-        setPopupState(function (state) {
-            return !state;
-        });
-    };
-
-    var buttonLabel = props.buttonLabel,
-        dateLabel = props.dateLabel,
-        date = props.date,
-        onChangeDate = props.onChangeDate,
-        onDateClear = props.onDateClear,
-        onInvalidDate = props.onInvalidDate;
-
-
-    return React.createElement(
-        Fragment,
-        null,
-        React.createElement(
-            "div",
-            { className: "advgb-advcalendar-control" },
-            React.createElement(
-                "label",
-                null,
-                dateLabel
-            ),
-            React.createElement(
-                "div",
-                null,
-                React.createElement(
-                    Button,
-                    {
-                        isLink: true,
-                        icon: "calendar",
-                        onClick: function onClick() {
-                            return setPopupState(togglePopup);
-                        }
-                    },
-                    React.createElement(
-                        Tooltip,
-                        { text: __('Change date', 'advanced-gutenberg') },
-                        React.createElement(
-                            "span",
-                            null,
-                            date ? moment(date).format("MMMM DD YYYY, h:mm a") : buttonLabel
-                        )
-                    )
-                ),
-                date && React.createElement(Button, {
-                    icon: "no-alt",
-                    className: "advgb-advcalendar-remove-icon",
-                    onClick: function onClick() {
-                        return onDateClear();
-                    }
-                })
-            )
-        ),
-        popupState && React.createElement(
-            Popover,
-            {
-                className: "advgb-advcalendar-popover",
-                onClose: setPopupState.bind(null, false)
-            },
-            React.createElement(
-                "label",
-                { className: "advgb-advcalendar-popover-label" },
-                dateLabel,
-                React.createElement(Button, {
-                    icon: "no-alt",
-                    className: "advgb-advcalendar-remove-icon",
-                    onClick: function onClick() {
-                        return setPopupState(togglePopup);
-                    }
-                })
-            ),
-            React.createElement(
-                "div",
-                { className: "advgb-advcalendar-popover-timezone" },
-                typeof advgbBlocks.timezone !== 'undefined' && advgbBlocks.timezone.length ? advgbBlocks.timezone.replace(/_/g, ' ') + " " + __('time', 'advanced-gutenberg') : __('WordPress settings timezone', 'advanced-gutenberg')
-            ),
-            React.createElement(DateTimePicker, {
-                currentDate: date,
-                onChange: onChangeDate,
-                is12Hour: true,
-                isInvalidDate: onInvalidDate
-            })
-        )
     );
 }
